Refactor text2sql based on ERAG (#1080)
