--- comps/texttosql/README.md
@@ -1,21 +1,0 @@
-# 🛢 Text-to-SQL Microservice
-
-In today's data-driven world, the ability to efficiently extract insights from databases is crucial. However, querying databases often requires specialized knowledge of SQL(Structured Query Language) and database schemas, which can be a barrier for non-technical users. This is where the Text-to-SQL microservice comes into play, leveraging the power of LLMs and agentic frameworks to bridge the gap between human language and database queries. This microservice is built on LangChain/LangGraph frameworks.
-
-The microservice enables a wide range of use cases, making it a versatile tool for businesses, researchers, and individuals alike. Users can generate queries based on natural language questions, enabling them to quickly retrieve relevant data from their databases. Additionally, the service can be integrated into ChatBots, allowing for natural language interactions and providing accurate responses based on the underlying data. Furthermore, it can be utilized to build custom dashboards, enabling users to visualize and analyze insights based on their specific requirements, all through the power of natural language.
-
----
-
-## 🛠️ Features
-
-**Implement SQL Query based on input text**: Transform user-provided natural language into SQL queries, subsequently executing them to retrieve data from SQL databases.
-
----
-
-## ⚙️ Implementation
-
-The text-to-sql microservice able to implement with various framework and support various types of SQL databases.
-
-### 🔗 Utilizing Text-to-SQL with Langchain framework
-
-For additional information, please refer to this [README](./langchain/README.md)
